Updated HTML/PHP pages
I formatted the HTML so it will be easier to use with the PHP file - especially the jQuery one. I also added in a contact.php based off the file Hiro sent us, but I am not sure how it is supposed to be implemented....We need to meet to solidify this code (the 01 and 02 assignments).