Updated Files
I added in the specific contact.php code Hiro wants us to use for this assignment to the php files. We just need to set up the form to speak to this file. I also updated the html files (for all 3 assignments) so that the input field names match what is in the php file we have been given. I also added in a "textarea" input field.