Worked part of AXIOS files
We attempted some AXIOS stuff, we have the js on the index.html page set up to sent post data to the server.js file we created. We just need to get our js file to do stuff with it and there's not a lot of helpful information out there to show us how...Austin is gonna work on it later today and get back to us.